В отдельный модуль выделена инъекция зависимостей для data-слоя
