+++ app/build.gradle
@@ -38,29 +38,20 @@ android {
         release {
             minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-            buildConfigField "String", "BASE_URL", '"https://test.vhospice.org/api/fmh/"'
-        }
-        debug {
-            buildConfigField "String", "BASE_URL", '"https://test.vhospice.org/api/fmh/"'
         }
     }
 }
 
 dependencies {
     implementation(project(":data"))
+    implementation(project(":data:di"))
+    implementation(project(":model"))
 
-    def retrofit_version = "2.9.0"
-    def retrofit_gson_version = "2.9.0"
-    def okhttp_logging_version = "4.9.2"
     def fragment_version = "1.4.1"
     def lifecycle_version = "2.4.1"
-    def room_version = '2.4.2'
     def recycler_version = '1.2.1'
     def material_version = '1.5.0'
     def constraint_version = '2.1.3'
-    def coroutines_version = '1.5.2'
-    def hilt_version = '2.41'
-    def gson_version = '2.9.0'
     def swipe_refresh_version = '1.1.0'
 
     implementation 'androidx.core:core-ktx:1.7.0'
@@ -68,13 +59,8 @@ dependencies {
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:$swipe_refresh_version"
     implementation "com.google.android.material:material:$material_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraint_version"
-    implementation "androidx.room:room-runtime:$room_version"
-    kapt "androidx.room:room-compiler:$room_version"
-    implementation "androidx.room:room-ktx:$room_version"
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
-    implementation "com.squareup.retrofit2:converter-gson:$retrofit_gson_version"
     implementation "com.google.code.gson:gson:$gson_version"
-    implementation "com.squareup.okhttp3:logging-interceptor:$okhttp_logging_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:$coroutines_version"
     implementation "androidx.fragment:fragment-ktx:$fragment_version"
@@ -91,8 +77,12 @@ dependencies {
 
     implementation "androidx.recyclerview:recyclerview:$recycler_version"
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:1.1.5")
+
+    // region Hilt
     implementation "com.google.dagger:hilt-android:$hilt_version"
     kapt "com.google.dagger:hilt-compiler:$hilt_version"
+    // endregion Hilt
+
     implementation platform('com.google.firebase:firebase-bom:29.0.3')
     implementation 'com.google.firebase:firebase-analytics-ktx'
     implementation 'com.google.firebase:firebase-crashlytics-ktx'
