--- app/src/main/java/ru/iteco/fmhandroid/api/qualifier/Refresh.kt
@@ -1,7 +1,0 @@
-package ru.iteco.fmhandroid.api.qualifier
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class Refresh
