--- app/src/main/java/ru/iteco/fmhandroid/db/DbModule.kt
@@ -1,21 +1,0 @@
-package ru.iteco.fmhandroid.db
-
-import android.content.Context
-import androidx.room.Room
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object DbModule {
-    @Provides
-    @Singleton
-    fun provideDb(@ApplicationContext context: Context): AppDb {
-        return Room.databaseBuilder(context, AppDb::class.java, "app.db")
-            .build()
-    }
-}
