+++ app/src/main/java/ru/iteco/fmhandroid/dto/NewsFilterArgs.kt
@@ -1,0 +1,10 @@
+package ru.iteco.fmhandroid.dto
+
+import android.os.Parcelable
+
+@kotlinx.parcelize.Parcelize
+data class NewsFilterArgs(
+    val category: String? = null,
+    val dates: List<Long>? = null,
+    val status: Boolean? = null
+) : Parcelable
