--- app/src/main/java/ru/iteco/fmhandroid/repository/authRepository/AuthRepositoryModule.kt
@@ -1,14 +1,0 @@
-package ru.iteco.fmhandroid.repository.authRepository
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import ru.iteco.fmh.data.AuthRepository
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class AuthRepositoryModule {
-    @Binds
-    abstract fun bindAuthRepository(impl: AuthRepositoryImpl): AuthRepository
-}
