--- app/src/main/java/ru/iteco/fmhandroid/repository/newsRepository/NewsRepositoryModule.kt
@@ -1,15 +1,0 @@
-package ru.iteco.fmhandroid.repository.newsRepository
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class NewsRepositoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindNewsRepository(impl: NewsRepositoryImpl): NewsRepository
-}
