--- app/src/main/java/ru/iteco/fmhandroid/repository/userRepository/UserRepositoryModule.kt
@@ -1,15 +1,0 @@
-package ru.iteco.fmhandroid.repository.userRepository
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-abstract class UserRepositoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindUserRepository(impl: UserRepositoryImpl): UserRepository
-}
