+++ app/src/main/java/ru/iteco/fmhandroid/ui/CreateEditClaimFragment.kt
@@ -20,7 +20,7 @@ import kotlinx.coroutines.launch
 import ru.iteco.fmhandroid.R
 import ru.iteco.fmhandroid.databinding.FragmentCreateEditClaimBinding
 import ru.iteco.fmhandroid.dto.Claim
-import ru.iteco.fmhandroid.dto.User
+import ru.iteco.fmh.model.User
 import ru.iteco.fmhandroid.utils.Utils
 import ru.iteco.fmhandroid.utils.Utils.fullUserNameGenerator
 import ru.iteco.fmhandroid.utils.Utils.saveDateTime
