+++ build.gradle
@@ -4,6 +4,11 @@ buildscript {
         kotlin_version = "1.6.10"
         hilt_version = "2.40.5"
         nav_version = "2.4.2"
+        hilt_version = '2.41'
+        retrofit_version = "2.9.0"
+        gson_version = '2.9.0'
+        coroutines_version = '1.5.2'
+        room_version = '2.4.2'
     }
 
     repositories {
@@ -12,7 +17,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:7.2.0'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.21"
         classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
         classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
         classpath 'com.google.gms:google-services:4.3.10'
