+++ data/di/src/main/kotlin/ru/iteco/fmh/data/di/db/DbModule.kt
@@ -1,0 +1,23 @@
+package ru.iteco.fmh.data.di.db
+
+import android.content.Context
+import androidx.room.Room
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import ru.iteco.fmh.data.db.AppDb
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object DbModule {
+    @Provides
+    @Singleton
+    fun provideDb(@ApplicationContext context: Context): AppDb {
+        return Room.databaseBuilder(
+            context, AppDb::class.java, "app.db"
+        ).build()
+    }
+}
