+++ data/di/src/main/kotlin/ru/iteco/fmh/data/di/qualifier/Refresh.kt
@@ -1,0 +1,7 @@
+package ru.iteco.fmh.data.di.qualifier
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+internal annotation class Refresh
