+++ data/di/src/main/kotlin/ru/iteco/fmh/data/di/repository/NewsRepositoryModule.kt
@@ -1,0 +1,17 @@
+package ru.iteco.fmh.data.di.repository
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import ru.iteco.fmh.data.NewsRepository
+import ru.iteco.fmh.data.impl.NewsRepositoryImpl
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+abstract class NewsRepositoryModule {
+    @Binds
+    @Singleton
+    abstract fun bindNewsRepository(impl: NewsRepositoryImpl): NewsRepository
+}
